test: add tests for reusable hook
